Create the process for CHI docker environment

+++ scripts/docker.sh
@@ -21,6 +21,7 @@ fi
 cp ${REPO_PATH}/Dockerfile ${TMPDIR}/
 cp ${REPO_PATH}/package.json ${TMPDIR}/package_chi.json
 cp ${REPO_PATH}/src/custom-elements/package.json ${TMPDIR}/package_custom-elements.json
+cp ${REPO_PATH}/src/chi-vue/package.json ${TMPDIR}/package_vue.json
 cp ${REPO_PATH}/scripts/entrypoint.sh ${TMPDIR}/
 
 pushd "${TMPDIR}"  > /dev/null
@@ -40,12 +41,18 @@ then
   mkdir ${REPO_PATH}/src/custom-elements/node_modules 2>/dev/null || die "${REPO_PATH}/src/custom-elements/node_modules must exists and be a directory"
 fi
 
+if [ ! -d "${REPO_PATH}/src/chi-vue/node_modules" ]
+then
+  mkdir ${REPO_PATH}/src/chi-vue/node_modules 2>/dev/null || die "${REPO_PATH}/src/chi-vue/node_modules must exists and be a directory"
+fi
+
 if [ "x$1" = "xstart" ]; then
 
   docker run --rm -it --name chi \
           --privileged \
           -v ${REPO_PATH}:/chi \
           -p 8000:8000 \
+          -p 9090:9090 \
           -p 8001:8001 \
           -p 3333:3333 \
           chi
